requirement.txt
annotated-types   ==  0.6.0
anyio             == 4.2.0
Brotli            == 1.1.0
certifi           == 2024.2.2
charset-normalizer== 3.3.2
fastapi           == 0.109.2
filelock          == 3.13.1
fsspec            == 2024.2.0
idna              == 3.6
Jinja2            == 3.1.3
MarkupSafe        == 2.1.5
mpmath            == 1.3.0
networkx          == 3.2.1
numpy             == 1.26.4
pillow            == 10.2.0
pip               == 24.0
pydantic          == 2.6.1
pydantic_core     == 2.16.2
PySocks           == 1.7.1
PyYAML            == 6.0.1
requests          == 2.31.0
setuptools        == 69.1.0
sniffio           == 1.3.0
starlette         == 0.36.3
sympy             == 1.12
torch             == 2.2.0
torchsummary      == 1.5.1
torchvision       == 0.17.0
tqdm              == 4.66.2
typing_extensions == 4.9.0
urllib3           == 2.2.1
wheel             == 0.42.0